\name{ZeligBLME-package}
\alias{ZeligBLME-package}
\alias{ZeligBLME}
\docType{package}
\title{
Add blme library to Zelig 4.
}
\description{
Longer description
}
\details{
\tabular{ll}{
Package: \tab ZeligBLME\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-09\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Benjamin Soltoff <ben@bensoltoff.com>
}
\references{
}
\keyword{ ZeligBLME }
\seealso{
}
\examples{
}
